Water Website Internet File (Not Application).
This is a internet file that takes you directly to the water website, likewise every website practically on the internet! This file is an internet viewing file and will simply just take the website into your own web browser. This can work on any OS.

However, it is important to include that this is not the formal application of the website. It can only work on MacOS, but this can work on any OS, since it is an internet viewing file.